+++ CHANGELOG.rst
@@ -236,6 +236,9 @@ New features:
   just for the current user. To enable a specific writer (if present in a given project) without enabling all writers of the same 
   class, set the ``alias=`` attribute on the writer configuration and specify the writer.
 
+- Added ``artifacts`` dictionary to `pysys.writer.outcomes.JSONResultsWriter` which records artifact paths published during execution of the tests, for 
+  example code coverage and performance reports.
+
 - Added environment variable ``PYSYS_PROJECT_APPEND`` which treats the main ``pysysproject.xml`` file as if it had additional XML from 
   the specified file appended to it. This allows user-specific settings such as custom writers to be added dynamically without 
   modifying the main project file. The XML file to append must have ``<pysysproject>`` as its root XML node.